+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 3.2.1
+* Properly handle macOS newlines in docblocks
 * Fix docblocks containing tabs instead of spaces not processing properly
 
 ## 3.2.0